web_server/requirements.txt
requests
azure-storage-blob
Flask
